requirements.txt
pandas>=1.0
numpy>=1.18
matplotlib>=3.1
mpld3>=0.5.9
colour-science>=0.3.16
tk
